scripts/requirements_send.txt
Pillow==10.1.0
requests==2.31.0
matplotlib==3.8.0
